False night trigger fix
Added code to speed up the timer if a false night (low light level) was
detected such as a cloud, solar eclipse, or bird standing on the
sensor, etc… This will shut the LEDs of faster. If the light level
drops below the trigger level again the timer will slow back down.